Lots of cross platform work.

+++ include/libnitro.h
@@ -1,10 +1,10 @@
 #ifndef LIBNITRO_H
 #define LIBNITRO_H
 
-// Opaque handling for holding state.
+/* Opaque handling for holding state. */
 typedef struct NitroCopyState NitroCopyState;
 
-// Define status codes.
+/* Define status codes. */
 typedef enum {
     NITRO_SUCCESS = 0,
     NITRO_ERROR_OPENING_FILE,
@@ -17,18 +17,18 @@ typedef enum {
 } NitroStatus;
     
 
-// State management functions.
-NitroCopyState* nitro_init(int overwrite);
+/* State management functions. */
+NitroCopyState* nitro_init(unsigned int overwrite);
 void nitro_destroy(NitroCopyState* state);
 
-// Core functionality for copying files and directories.
+/* Core functionality for copying files and directories. */
 NitroStatus nitro_copy(NitroCopyState* state, const char* src, const char* dest);
 NitroStatus nitro_copy_file(NitroCopyState* state, const char* src, const char* dest);
 NitroStatus nitro_copy_directory(NitroCopyState* state, const char* src, const char* dest);
 
-// Utility functions.
-NitroStatus nitro_get_total_stats(const char* path, long long* total_size, long long* file_count);
+/* Utility functions. */
+NitroStatus nitro_get_total_stats(const char* path, unsigned long* total_size, unsigned long* file_count);
 void nitro_update_progress(NitroCopyState* state);
-char* nitro_format_bytes(long long bytes);
+char* nitro_format_bytes(unsigned long bytes);
 
 #endif
